requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tree-sitter==0.20.4
tree-sitter-python==0.21.0
tree-sitter-go==0.21.0
tree-sitter-java==0.21.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
psycopg2-binary==2.9.7
pymongo==4.5.0
python-dotenv==1.0.0
pydantic==2.4.2
loguru==0.7.2
gitpython==3.1.40
pytest==7.4.3
httpx==0.25.1 